refactor: extract TaskEditInArchiveDialog into its own component file
Move the task-editing dialog from ArchiveEditDialog.tsx into a dedicated
TaskEditInArchiveDialog.tsx, reducing ArchiveEditDialog from 850 to ~270
lines. Unused Select imports are removed from ArchiveEditDialog.
